Move file list into component
