+++ assets/src/scripts/results-viewer/App.jsx
@@ -1,6 +1,7 @@
 import React from "react";
-import { QueryClient, QueryClientProvider, useQuery } from "react-query";
+import { QueryClient, QueryClientProvider } from "react-query";
 import { ReactQueryDevtools } from "react-query/devtools";
+import FileList from "./FileList";
 
 const queryClient = new QueryClient({
   defaultOptions: {
@@ -9,27 +10,6 @@ const queryClient = new QueryClient({
     },
   },
 });
-
-function FileList() {
-  const { isLoading, isError, data, error } = useQuery(
-    "FILE_LIST",
-    () => fetch("/"),
-    {
-      initialData: window.props,
-      staleTime: Infinity,
-    }
-  );
-
-  if (isLoading) {
-    return <span>Loading...</span>;
-  }
-
-  if (isError) {
-    return <span>Error: {error.message}</span>;
-  }
-
-  return <h1>{data.name}</h1>;
-}
 
 function App() {
   return (
