+++ assets/src/scripts/results-viewer/FileList.jsx
@@ -1,0 +1,39 @@
+import React from "react";
+import { useQuery } from "react-query";
+
+const dummyData = [
+  "src/file1.html",
+  "src/file2.svg",
+  "src/file3.csv",
+  "src/file4.png",
+  "src/file5.txt",
+];
+
+function FileList() {
+  const { isLoading, isError, data, error } = useQuery(
+    "FILE_LIST",
+    () => fetch("/"),
+    {
+      initialData: dummyData,
+      staleTime: Infinity,
+    }
+  );
+
+  if (isLoading) {
+    return <span>Loading...</span>;
+  }
+
+  if (isError) {
+    return <span>Error: {error.message}</span>;
+  }
+
+  return (
+    <ul>
+      {data.map((item) => (
+        <li key={item}>{item}</li>
+      ))}
+    </ul>
+  );
+}
+
+export default FileList;
